Resolves merge conflicts between APY integration and TVL/debts migration.
This merge combines both features to source TVL, debts, and APY data from
Kong's GraphQL API as a single source of truth.

Key changes:
- Combined KongVault struct to include TVL, debts, and APY fields
- Updated GraphQL queries to fetch all three data types in single request
- Added unified storage layer with GetKongTVL(), GetKongDebts(), GetKongAPY()
- Fixed type conversions: Kong TVL (float64) and debts (BigInt as string)
- Simplified APR calculation processes to use Kong APY data
- Removed duplicate GetKongAPY function from storage layer

Files modified:
- internal/kong/client.go: Added APY to KongVault, updated queries
- internal/models/vaults.go: Added KongAPY struct to TKongVaultSchema
- internal/storage/elem.vaults.go: Added APY getter, removed duplicate
- internal/fetcher/builder.go: Uses Kong TVL with float64 conversion
- processes/apr/*.go: Refactored to use Kong APY data

+++ internal/fetcher/builder.go
@@ -186,7 +186,7 @@ func BuildVaultTVL(t models.TVault) models.TTVL {
 
 	tvl := models.TTVL{
 		TotalAssets: t.LastTotalAssets,
-		TVL:		float64(kongTVL),
+		TVL:         kongTVL,
 		Price:       fHumanizedPrice,
 	}
 	return tvl
